ძებნები ახალი

+++ src/main/java/com/george/orca/service/LoanPaymentService.java
@@ -22,8 +22,8 @@ public interface LoanPaymentService {
                                  String debtor,
                                  String debtorIdentificator,
                                  String assignedAgent,
-                                 BigDecimal amountStart,
-                                 BigDecimal amountEnd,
+                                 Long amountStart,
+                                 Long amountEnd,
                                  String dateStart,
                                  String dateEnd);
